requirements.txt
alembic==1.0.9
Click==7.0
dlib==19.17.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
